+++ src/resources/Locale.ts
@@ -1,5 +1,6 @@
-export default {
-    //format 1
+//data
+
+const fromData = {
     "ar-AE": "Arabic_UAE",
     "de-DE": "German_Germany",
     "en-GB": "English_United_Kingdom", //not support valorant-api.com
@@ -19,27 +20,45 @@ export default {
     "vi-VN": "Vietnamese_Vietnam",
     "zh-CN": "Chinese_China",
     "zh-TW": "Chinese_Taiwan",
+};
 
-    data: {
-        //format 2
-        Arabic_UAE: "ar-AE",
-        German_Germany: "de-DE",
-        English_United_Kingdom: "en-GB",
-        English_United_States: "en-US",
-        Spanish_Spain: "es-ES",
-        Spanish_Mexico: "es-MX",
-        French_France: "fr-FR",
-        Indonesian_Indonesia: "id-ID",
-        Italian_Italy: "it-IT",
-        Japanese_Japan: "ja-JP",
-        Korean_Korea: "ko-KR",
-        Polish_Poland: "pl-PL",
-        Portuguese_Brazil: "pt-BR",
-        Russian_Russia: "ru-RU",
-        Thai_Thailand: "th-TH",
-        Turkish_Turkey: "tr-TR",
-        Vietnamese_Vietnam: "vi-VN",
-        Chinese_China: "zh-CN",
-        Chinese_Taiwan: "zh-TW",
-    },
+const toData = {
+    Arabic_UAE: "ar-AE",
+    German_Germany: "de-DE",
+    English_United_Kingdom: "en-GB",
+    English_United_States: "en-US",
+    Spanish_Spain: "es-ES",
+    Spanish_Mexico: "es-MX",
+    French_France: "fr-FR",
+    Indonesian_Indonesia: "id-ID",
+    Italian_Italy: "it-IT",
+    Japanese_Japan: "ja-JP",
+    Korean_Korea: "ko-KR",
+    Polish_Poland: "pl-PL",
+    Portuguese_Brazil: "pt-BR",
+    Russian_Russia: "ru-RU",
+    Thai_Thailand: "th-TH",
+    Turkish_Turkey: "tr-TR",
+    Vietnamese_Vietnam: "vi-VN",
+    Chinese_China: "zh-CN",
+    Chinese_Taiwan: "zh-TW",
+};
+
+//exchange
+
+function toString(x: keyof typeof toData): keyof typeof fromData {
+    return toData[x] as keyof typeof fromData;
+}
+
+function fromString(x: keyof typeof fromData): keyof typeof toData {
+    return fromData[x] as keyof typeof toData;
+}
+
+//export
+
+export default {
+    from: fromData,
+    to: toData,
+    toString,
+    fromString,
 };