+++ src/resources/QueueId.ts
@@ -1,5 +1,4 @@
-export default {
-    //format 1
+const fromData = {
     'unrated': 'Unrated',
     'competitive': 'Competitive',
     'spikerush': 'Spikerush',
@@ -11,16 +10,34 @@ export default {
 
     //change when have new map
     'newmap': 'Fracture',
+};
 
-    data: {
-        //format 2
-        Unrated: 'unrated',
-        Competitive: 'competitive',
-        Spikerush: 'spikerush',
-        Deathmatch: 'deathmatch',
-        Escalation: 'ggteam',
-        Replication: 'onefa',
-        Snowball_Fight: 'snowball',
-        Custom: '',
-    },
+const toData = {
+    Unrated: 'unrated',
+    Competitive: 'competitive',
+    Spikerush: 'spikerush',
+    Deathmatch: 'deathmatch',
+    Escalation: 'ggteam',
+    Replication: 'onefa',
+    Snowball_Fight: 'snowball',
+    Custom: '',
+};
+
+//exchange
+
+function toString(x: keyof typeof toData): keyof typeof fromData {
+    return toData[x] as keyof typeof fromData;
+}
+
+function fromString(x: keyof typeof fromData): keyof typeof toData {
+    return fromData[x] as keyof typeof toData;
+}
+
+//export
+
+export default {
+    from: fromData,
+    to: toData,
+    toString,
+    fromString,
 };