+++ src/resources/Region.ts
@@ -1,5 +1,4 @@
-export default {
-    //format 1
+const fromData = {
     'na': 'North_America',
     'latam': 'Latin_America',
     'br': 'Brazil',
@@ -7,15 +6,33 @@ export default {
     'eu': 'Europe',
     'kr': 'Korea',
     'ap': 'Asia_Pacific',
-    
-    data: {
-        //format 2
-        'North_America': 'na',
-        'Latin_America': 'latam',
-        'Brazil': 'br',
-        'Public_Beta_Environment': 'pbe',
-        'Europe': 'eu',
-        'Korea': 'kr',
-        'Asia_Pacific': 'ap',
-    },
+};
+
+const toData = {
+    North_America: 'na',
+    Latin_America: 'latam',
+    Brazil: 'br',
+    Public_Beta_Environment: 'pbe',
+    Europe: 'eu',
+    Korea: 'kr',
+    Asia_Pacific: 'ap',
+};
+
+//exchange
+
+function toString(x: keyof typeof toData): keyof typeof fromData {
+    return toData[x] as keyof typeof fromData;
+}
+
+function fromString(x: keyof typeof fromData): keyof typeof toData {
+    return fromData[x] as keyof typeof toData;
+}
+
+//export
+
+export default {
+    from: fromData,
+    to: toData,
+    toString,
+    fromString,
 };