+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# 1.2.0 (28 May, 2020)
+
+- big update to **`resources`**
+    - ~~*{name}*~~ *==>* {name}**.to**
+    - ~~*{name}*.data~~ *==>* {name}**.from**
+    - {name}**.toString()**
+    - {name}**.fromString()**
+- ~~CustomEvent~~ *==>* **ValEvent**
+- ~~AxiosClient~~ *==>* **ValRequestClient**
+    - Move 4 AxiosClient from all modules to only one module (this module)
+- **`.toUft8()`**
+    - remove repeated code
+
 # 1.0.0 (1 January, 2020)
 
 - Initial release